feat: fix checkbox
