+++ src/components/ui/custom-toggle.tsx
@@ -1,0 +1,7 @@
+import type { ComponentProps } from 'react';
+
+import { Toggle as ShadToggle } from '@/components/ui/toggle';
+
+export function Toggle(props: ComponentProps<typeof ShadToggle>) {
+  return <ShadToggle {...props} />;
+}
