+++ src/feature/auth/registration/registration-schema.tsx
@@ -51,5 +51,4 @@ export const schema = object({
   streetName: nameValidator,
   setAsDefaultShipping: boolean(),
   setAsDefaultBilling: boolean(),
-  useSameForBilling: boolean(),
 });
